move NewsItem struct to DMONewsProfile
